requirements.txt
Flask==2.0.3
language-tool-python==2.7.1
pyspellchecker==0.6.2
